Changed the MovieRepository to an Interface with Implementation
